+++ .gitignore
@@ -1,5 +1,5 @@
 .class
 lib/
 .idea
-JB_2021_29.iml
+src/otus_java_basic/Marchenko/JB_2021_29.iml
 out
