--- src/Exception/AlphabetEntryException.java
@@ -1,7 +1,0 @@
-package Exception;
-
-public class AlphabetEntryException extends RuntimeException {
-    public AlphabetEntryException(String massage) {
-        super(massage);
-    }
-}
