--- src/ProjectOOP/Answer.java
@@ -1,14 +1,0 @@
-package ProjectOOP;
-
-public class Answer {
-
-    private int correctAnswerIndex;
-
-    public Answer (int correctAnswerIndex) {
-        this.correctAnswerIndex = correctAnswerIndex;
-    }
-
-    public int getCorrectAnswerIndex() {
-        return correctAnswerIndex;
-    }
-}
