--- src/ProjectOOP/Question.java
@@ -1,19 +1,0 @@
-package ProjectOOP;
-
-public class Question {
-    private String description;
-    private Answer answer;
-
-    public Question(String description, Answer answer){
-        this.answer = answer;
-        this.description = description;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public Answer getAnswer() {
-        return answer;
-    }
-}
