--- src/Testing/App/Demo.java
@@ -1,13 +1,0 @@
-package Testing.App;
-
-import Testing.App.Game.*;
-
-public class Demo {
-
-    public static void main(String[] args) {
-        Dice dice = new DiceImpl();
-        GameWinnerPrinter winnerPrinter = new GameWinnerConsolePrinter();
-        Game game = new Game(dice, winnerPrinter);
-        game.playGame(new Player("Вася"), new Player("Игорь"));
-    }
-}
