--- src/Testing/App/Game/Dice.java
@@ -1,5 +1,0 @@
-package Testing.App.Game;
-
-public interface Dice {
-    int roll();
-}
