--- src/Testing/App/Game/GameWinnerConsolePrinter.java
@@ -1,8 +1,0 @@
-package Testing.App.Game;
-
-public class GameWinnerConsolePrinter implements GameWinnerPrinter{
-    @Override
-    public void printWinner(Player winner) {
-        System.out.printf("Победитель: %s%n", winner.getName());
-    }
-}
