--- src/Testing/App/Game/GameWinnerPrinter.java
@@ -1,5 +1,0 @@
-package Testing.App.Game;
-
-public interface GameWinnerPrinter {
-    void printWinner(Player winner);
-}
