--- src/Testing/App/Game/Player.java
@@ -1,13 +1,0 @@
-package Testing.App.Game;
-
-public class Player {
-    private final String name;
-
-    public Player(String name) {
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-}
