+++ src/otus_java_basic/Marchenko/Exception/AlphabetEntryException.java
@@ -1,0 +1,7 @@
+package otus_java_basic.Marchenko.Exception;
+
+public class AlphabetEntryException extends RuntimeException {
+    public AlphabetEntryException(String massage) {
+        super(massage);
+    }
+}
