+++ src/otus_java_basic/Marchenko/FirstProgram/HomeWork.java
@@ -1,0 +1,14 @@
+package otus_java_basic.Marchenko.FirstProgram;
+import java.util.Base64;
+import java.util.Scanner;
+
+public class HomeWork {
+    
+    public static void main(String[] args) {
+        Scanner sc = new Scanner(System.in);
+        System.out.println("Please enter you full name: ");
+        String fullName = sc.nextLine();
+        String encodedFullName = Base64.getEncoder().encodeToString(fullName.getBytes());
+        System.out.println("Program result: " + encodedFullName);
+    }
+}
