+++ src/otus_java_basic/Marchenko/ProjectOOP/Answer.java
@@ -1,0 +1,14 @@
+package otus_java_basic.Marchenko.ProjectOOP;
+
+public class Answer {
+
+    private int correctAnswerIndex;
+
+    public Answer (int correctAnswerIndex) {
+        this.correctAnswerIndex = correctAnswerIndex;
+    }
+
+    public int getCorrectAnswerIndex() {
+        return correctAnswerIndex;
+    }
+}
