+++ src/otus_java_basic/Marchenko/Testing/App/Demo.java
@@ -1,0 +1,13 @@
+package otus_java_basic.Marchenko.Testing.App;
+
+import otus_java_basic.Marchenko.Testing.App.Game.*;
+
+public class Demo {
+
+    public static void main(String[] args) {
+        Dice dice = new DiceImpl();
+        GameWinnerPrinter winnerPrinter = new GameWinnerConsolePrinter();
+        Game game = new Game(dice, winnerPrinter);
+        game.playGame(new Player("Вася"), new Player("Игорь"));
+    }
+}
