+++ src/otus_java_basic/Marchenko/Testing/App/Game/Dice.java
@@ -1,0 +1,5 @@
+package otus_java_basic.Marchenko.Testing.App.Game;
+
+public interface Dice {
+    int roll();
+}
