+++ src/otus_java_basic/Marchenko/Testing/App/Game/DiceImpl.java
@@ -1,0 +1,11 @@
+package otus_java_basic.Marchenko.Testing.App.Game;
+
+import java.util.Random;
+
+public class DiceImpl implements Dice{
+    @Override
+    public int roll() {
+        // https://docs.oracle.com/javase/8/docs/api/java/util/Random.html
+        return new Random().nextInt();
+    }
+}
