+++ src/otus_java_basic/Marchenko/Testing/App/Game/GameWinnerPrinter.java
@@ -1,0 +1,5 @@
+package otus_java_basic.Marchenko.Testing.App.Game;
+
+public interface GameWinnerPrinter {
+    void printWinner(Player winner);
+}
