+++ src/otus_java_basic/Marchenko/Testing/App/Game/Player.java
@@ -1,0 +1,13 @@
+package otus_java_basic.Marchenko.Testing.App.Game;
+
+public class Player {
+    private final String name;
+
+    public Player(String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
