+++ src/otus_java_basic/Marchenko/otus_test.sql
